Analysis of Covid-19 Containment Measures
This Python file contains an in-depth analysis of the effectiveness of Covid-19 containment measures implemented globally compared to those specifically in the United States. The analysis utilizes data from 2020 through 2021, sourced from publicly available WHO data. It includes visualizations, statistical insights, and recommendations based on the findings. This analysis aims to provide valuable insights into the impact of different strategies on controlling the spread of the virus.