Effectiveness of Covid-19 Containment Measures
This PowerPoint presentation complements the in-depth analysis of the effectiveness of Covid-19 containment measures on a global scale, with a specific focus on the United States. The analysis, based on data spanning from 2020 to 2021 and sourced from publicly available WHO data, encompasses comprehensive visualizations, statistical insights, and actionable recommendations. This presentation aims to provide a clear understanding of the impact of different strategies on controlling the spread of the virus, offering valuable insights for informed decision-making and policy formulation.